Pull request #4015: [BLUETOOTH-17655][LIB] HOST-FSCI: Fix MISRA C 2012 Rule 10.3 violations in fsci_ble_gap2_handlers.c
Merge in CONNINT/bluetooth_public from BLUETOOTH-17655-kw45-kw47-gap-le_modulation_test-hci-vendor-command-in-ble_test_appnew to develop/main

Squashed commit of the following:

commit c7081b0d862e9c488e48d2c94df775e1585be2cb
Author: Alexandru Vrabete <alexandru.vrabete@nxp.com>
Date:   Mon May 11 11:13:50 2026 +0300

    [BLUETOOTH-17655] HOST-FSCI: Fix MISRA C 2012 Rule 10.3 in fsci_ble_gap2_handlers.c

+++ fsci/source/fsci_ble_gap2_handlers.c
@@ -997,7 +997,7 @@ static void HandleGapCmdLeChasConfig
     uint32_t fsciInterfaceId
 )
 {
-    int8_t   initialFilterSeed;
+    uint8_t   initialFilterSeed;
     uint8_t  filterThreshold;
     uint8_t  filterWeight;
     uint16_t measurementInterval;
@@ -1008,7 +1008,7 @@ static void HandleGapCmdLeChasConfig
     fsciBleGetUint8ValueFromBuffer(filterWeight, pBuffer);
     fsciBleGetUint16ValueFromBuffer(measurementInterval, pBuffer);
 
-    fsciBleGap2CallApiFunction(Gap_LeChasConfig(initialFilterSeed, 
+    fsciBleGap2CallApiFunction(Gap_LeChasConfig((int8_t)initialFilterSeed, 
                                                  filterThreshold, 
                                                  filterWeight, 
                                                  measurementInterval));
@@ -1037,7 +1037,7 @@ static void HandleGapCmdLeModulationTest
     uint8_t  testMode;
     uint8_t  channelNumber;
     uint8_t  options;
-    int8_t   txPower;
+    uint8_t   txPower;
 
     /* Get command parameters from buffer */
     fsciBleGetUint8ValueFromBuffer(testMode, pBuffer);
@@ -1048,7 +1048,7 @@ static void HandleGapCmdLeModulationTest
     fsciBleGap2CallApiFunction(Gap_LeModulationTest(testMode, 
                                                      channelNumber, 
                                                      options, 
-                                                     txPower));
+                                                     (int8_t)txPower));
 }
 
 /*! *********************************************************************************
